requirements.txt
discord.py
aiohttp
websockets==3.4
chardet
qrcode[pil]
vk
wolframalpha
imgurpython
youtube_dl
pynacl


